+++ geojson_pydantic/features.py
@@ -16,10 +16,10 @@ class Feature(GenericModel, Generic[Geom, Props]):
     """Feature Model"""
 
     type: str = Field("Feature", const=True)
-    geometry: Geom
+    geometry: Geom = None
     properties: Optional[Props]
     id: Optional[str]
-    bbox: Optional[BBox]
+    bbox: Optional[BBox] = None
 
     class Config:
         """TODO: document"""
@@ -32,6 +32,11 @@ class Feature(GenericModel, Generic[Geom, Props]):
         if hasattr(v, "__geo_interface__"):
             return v.__geo_interface__
         return v
+
+    @property
+    def __geo_interface__(self):
+        """GeoJSON-like protocol for geo-spatial (GIS) vector data."""
+        return self.dict()
 
 
 class FeatureCollection(GenericModel, Generic[Geom, Props]):
@@ -52,3 +57,8 @@ class FeatureCollection(GenericModel, Generic[Geom, Props]):
     def __getitem__(self, index):
         """get feature at a given index"""
         return self.features[index]
+
+    @property
+    def __geo_interface__(self):
+        """GeoJSON-like protocol for geo-spatial (GIS) vector data."""
+        return self.dict()