+++ CHANGELOG.md
@@ -5,8 +5,46 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [Unreleased]
 
+### Added
+- `.wkt` property for Geometry object
+    ```python
+    from geojson_pydantic.geometries import Point
+
+    Point(coordinates=(1, 2)).wkt
+    >> 'POINT (1.0 2.0)'
+    ```
+
+- `.exterior` and `.interiors` properties for `geojson_pydantic.geometries.Polygon` object.
+    ```python
+    from geojson_pydantic.geometries import Polygon
+    polygon = Polygon(
+        coordinates=[
+            [(0, 0), (0, 10), (10, 10), (10, 0), (0, 0)],
+            [(2, 2), (2, 4), (4, 4), (4, 2), (2, 2)],
+        ]
+    )
+    polygon.exterior
+    >> [(0.0, 0.0), (0.0, 10.0), (10.0, 10.0), (10.0, 0.0), (0.0, 0.0)]
+
+    list(polygon.interiors)
+    >> [[(2.0, 2.0), (2.0, 4.0), (4.0, 4.0), (4.0, 2.0), (2.0, 2.0)]]
+    ```
+
+- `__geo_interface__` to `geojson_pydantic.geometries.GeometryCollection` object
+- `__geo_interface__` to `geojson_pydantic.feature.Feature` and `geojson_pydantic.feature.FeatureCollection` object
+
+### Changed
+- switch to `pyproject.toml`
+- rename `geojson_pydantic.version` to `geojson_pydantic.__version__`
+
+### Fixed
+- changelog compare links
+
+## [0.3.4] - 2022-04-28
+
+- Fix optional geometry and bbox fields on `Feature`; allowing users to pass in `None` or even omit either field (author @moradology, https://github.com/developmentseed/geojson-pydantic/pull/56)
 - Fix `Polygon.from_bounds` to respect geojson specification and return counterclockwise linear ring (author @jmfee-usgs, https://github.com/developmentseed/geojson-pydantic/pull/49)
 
 ## [0.3.3] - 2022-03-04
@@ -131,3 +169,15 @@ Although the type file was added in `0.2.0` it wasn't included in the distribute
 
 ### Added
 - Initial Release
+
+[unreleased]: https://github.com/developmentseed/geojson-pydantic/compare/0.3.4...HEAD
+[0.3.4]: https://github.com/developmentseed/geojson-pydantic/compare/0.3.3...0.3.4
+[0.3.3]: https://github.com/developmentseed/geojson-pydantic/compare/0.3.2...0.3.3
+[0.3.2]: https://github.com/developmentseed/geojson-pydantic/compare/0.3.1...0.3.2
+[0.3.1]: https://github.com/developmentseed/geojson-pydantic/compare/0.3.0...0.3.1
+[0.3.0]: https://github.com/developmentseed/geojson-pydantic/compare/0.2.3...0.3.0
+[0.2.3]: https://github.com/developmentseed/geojson-pydantic/compare/0.2.2...0.2.3
+[0.2.2]: https://github.com/developmentseed/geojson-pydantic/compare/0.2.1...0.2.2
+[0.2.1]: https://github.com/developmentseed/geojson-pydantic/compare/0.2.0...0.2.1
+[0.2.0]: https://github.com/developmentseed/geojson-pydantic/compare/0.1.0...0.2.0
+[0.1.0]: https://github.com/developmentseed/geojson-pydantic/compare/005f3e57ad07272c99c54302decc63eec12175c9...0.1.0